pyproject.toml
# mvp 단계의 toml mvp완료후 수정예정

[project]
name = "business-card"
version = "0.1.0"
description = "MVP backend for business card service"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "django>=5.1",
    "djangorestframework>=3.15",
    "drf-spectacular>=0.27.0",
    "pydantic[email]>=2.7",
    "boto3>=1.35",
    "mypy-boto3-s3>=1.35",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "mypy>=1.10",
    "ruff>=0.14.11",
    "pre-commit>=4.5.0",
    "django-stubs>=5.2.8",
]

[tool.django-stubs]
django_settings_module = "config.settings"
strict_settings = false

[tool.pyright]
venvPath = "."
venv = ".venv"
reportMissingImports = "none"
reportMissingModuleSource = "none"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = [
    "E501",  # line length (formatter에게 맡김)
]


[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true

ignore_missing_imports = true
check_untyped_defs = false
disallow_untyped_defs = false
# disallow_any_return = false

explicit_package_bases = true
warn_unused_configs = true

exclude = [
    "manage.py",
    "main.py",
    "django_app/infra/.*",
    "django_app/signals/.*",
    "django_app/migrations/.*",
]

# 앱별 완화
[[tool.mypy.overrides]]
module = [
    "django_app.models.*",
    "django_app.views.*"
]
ignore_errors = true

[[tool.mypy.overrides]]
module = "django_app.services.*"
check_untyped_defs = true
